# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/app-category/tag/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app类别"
                ],
                "summary": "获取某个类型的app的所有支持的tag号，即版本号",
                "parameters": [
                    {
                        "description": "app创建",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppLittleCategoryTagReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AppLittleCategoryTagRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app-category/template": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app类别"
                ],
                "summary": "获取所有支持的App类别模板",
                "responses": {
                    "200": {
                        "description": "获取所有支持的App类别模板",
                        "schema": {
                            "$ref": "#/definitions/model.AppCategoryTmplRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AppCategoryTmplBase": {
            "type": "object",
            "properties": {
                "app_little_category": {
                    "description": "标题",
                    "type": "string"
                },
                "env": {
                    "description": "必备的环境变量",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "icon": {
                    "description": "图标",
                    "type": "string"
                },
                "is_need_volume": {
                    "description": "是否需要磁盘",
                    "type": "boolean"
                },
                "show_name": {
                    "description": "名称",
                    "type": "string"
                },
                "summary": {
                    "description": "简介",
                    "type": "string"
                }
            }
        },
        "model.AppCategoryTmplBaseArr": {
            "type": "object",
            "properties": {
                "apps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AppCategoryTmplBase"
                    }
                },
                "name": {
                    "type": "string"
                },
                "show_name": {
                    "type": "string"
                }
            }
        },
        "model.AppCategoryTmplRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AppCategoryTmplBaseArr"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.AppLittleCategoryTagBase": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                }
            }
        },
        "model.AppLittleCategoryTagReq": {
            "type": "object",
            "required": [
                "current",
                "little_category",
                "pageSize"
            ],
            "properties": {
                "current": {
                    "description": "当前第几页",
                    "type": "integer"
                },
                "little_category": {
                    "type": "string"
                },
                "pageSize": {
                    "description": "每页多少条",
                    "type": "integer"
                },
                "search": {
                    "description": "模糊搜索字段，可不传",
                    "type": "string"
                }
            }
        },
        "model.AppLittleCategoryTagRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.AppLittleCategoryTagBase"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.Tag": {
            "type": "object",
            "properties": {
                "app_little_category": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "library": {
                    "type": "string"
                },
                "modify_time": {
                    "type": "string"
                },
                "version": {
                    "description": "版本号",
                    "type": "string"
                }
            }
        },
...
    }
}